Adds a postinstall script that initializes bowerrc
If .bowerrc file does not exist in the installing
project, it'll instantiate one.

This implementation has one goofy blind spot: in the event that
the grandparent of your actual project directory happens to 
contain a "package.json" file, you're going to get a .bowerrc 
up there. Oops. This would be avoided if it were clear npm had
a way for an installing module to know where the root was.

Fixes #3